requirements.txt
fastapi
uvicorn
python-dotenv
openai
openai-agents
pydantic
numpy